ric-app-ml/requirements.txt
sctp
socket
matplotlib
numpy
Pillow